requirements.txt
tweepy
python-dotenv
feedparser
logging
textblob
spacy
transformers
